+++ app/assets/javascripts/section/main.jsx
@@ -1,13 +1,13 @@
 import MixpanelUtils from '../helpers/mixpanel_utils.jsx';
+import SectionPage from './SectionPage'
 
 export default function renderSectionMain(el) {
-  const SectionPage = window.shared.SectionPage;
 
   const serializedData = $('#serialized-data').data();
   MixpanelUtils.registerUser(serializedData.currentEducator);
   MixpanelUtils.track('PAGE_VISIT', { page_key: 'SECTION' });
 
-  window.ReactDOM.render(<SectionPage 
+  window.ReactDOM.render(<SectionPage
         students={serializedData.students}
         section={serializedData.section}
         educators={serializedData.educators}
